move all scripts in package
