+++ render_service_scripts/configure_viewer.py
