+++ render_service_scripts/find_scene_core.py
