+++ render_service_scripts/launch_core_render.py
