+++ render_service_scripts/unpack.py
@@ -11,7 +11,7 @@ shutil.register_unpack_format('7zip', ['.7z'], unpack_7zarchive)
 
 
 def unpack_scene():
-    for entry in os.scandir('.'):
+    for entry in os.scandir('..'):
         if entry.name.endswith('.zip') or entry.name.endswith('.7z'):
             logging.info('Unpack {}'.format(entry.name))
             shutil.unpack_archive(entry.name, '.')
